pyproject.toml
[project]
name = "eventpropagator"
version = "0.1.0"
description = ""
authors = [
    {name = "Jonas Streckis",email = "jonas.str@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
packages = [{include = "eventpropagator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.0"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
asyncio = "^4.0.0"

[tool.poetry.scripts]
propagator = "eventpropagator:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
testpaths = ["tests"]